pyproject.toml
[project]
name = "legal-saathi-document-advisor"
version = "2.0.0"
description = "AI-powered platform for analyzing legal documents with FastAPI backend and React frontend"
authors = [
    {name = "LegalSaathi Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "google-generativeai>=0.3.0",
    "google-cloud-translate>=3.12.0",
    "google-cloud-documentai>=2.20.0",
    "google-cloud-language>=2.10.0",
    "google-cloud-speech>=2.21.0",
    "google-cloud-texttospeech>=2.16.0",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.0",
    "psutil==6.1.0",
    "reportlab>=4.0.0",
    "python-docx>=1.1.0",
    "requests>=2.31.0",
     "groq>=0.31.0",
    "pydantic>=2.0.0",
]
requires-python = ">=3.9"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]
